feat: implement logging module and refactor logging initialization
